Added

+++ bin/lib.sh
@@ -7,8 +7,8 @@ function pipe_download_to_command() {
         cd $OPENSHIFT_DATA_DIR
     fi
 
-    rm -rf $1
-    wget http://parltrack.euwiki.org/dumps/$1 || exit 1
+    [ -z "$NOCLEAN" ] && rm -rf $1
+    [ -f "$1" ] || wget http://parltrack.euwiki.org/dumps/$1 || exit 1
 
     if [ -n "$OPENSHIFT_REPO_DIR" ]; then
         cd $OPENSHIFT_REPO_DIR
@@ -16,4 +16,5 @@ function pipe_download_to_command() {
 
     export DJANGO_SETTINGS_MODULE=memopol.settings
     unxz -c ${OPENSHIFT_DATA_DIR}$1 | $2
+    [ -z "$NOCLEAN" ] && rm -rf $1
 }
